pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-cicd-triage"
version = "0.1.0"
description = "OpenEnv CI/CD debugging environment with deterministic grading"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openenv-core[core,cli]>=0.2.3",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pyyaml>=6.0.2",
  "openai>=1.30.0",
]

[project.optional-dependencies]
server = ["uvicorn>=0.30.0"]
dev = ["pytest>=8.2.0", "httpx>=0.27.0"]

[project.scripts]
server = "openenv_cicd_triage.server.app:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["openenv_cicd_triage*"]

[tool.setuptools.package-data]
openenv_cicd_triage = ["data/*.yaml"]
